Refactor application structure to utilize new component paths, enhancing modularity and maintainability. Update imports in various pages to reflect new component locations and add new components: ContactForm, ContactInfo, MapEmbed, and Footer for improved user experience.
